Inject ForwardCompatibiliy in TestCase
