--- src/Symfony/Bridge/PhpUnit/Legacy/ForwardCompatTestTraitForV7.php
@@ -1,50 +1,0 @@
-<?php
-
-/*
- * This file is part of the Symfony package.
- *
- * (c) Fabien Potencier <fabien@symfony.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace Symfony\Bridge\PhpUnit\Legacy;
-
-use PHPUnit\Framework\MockObject\MockObject;
-
-/**
- * @internal
- */
-trait ForwardCompatTestTraitForV7
-{
-    use ForwardCompatTestTraitForV5;
-
-    /**
-     * @param string|string[] $originalClassName
-     */
-    protected function createMock($originalClassName): MockObject
-    {
-        return $this->getMockBuilder($originalClassName)
-            ->disableOriginalConstructor()
-            ->disableOriginalClone()
-            ->disableArgumentCloning()
-            ->disallowMockingUnknownTypes()
-            ->getMock();
-    }
-
-    /**
-     * @param string|string[] $originalClassName
-     * @param string[]        $methods
-     */
-    protected function createPartialMock($originalClassName, array $methods): MockObject
-    {
-        return $this->getMockBuilder($originalClassName)
-            ->disableOriginalConstructor()
-            ->disableOriginalClone()
-            ->disableArgumentCloning()
-            ->disallowMockingUnknownTypes()
-            ->setMethods(empty($methods) ? null : $methods)
-            ->getMock();
-    }
-}
